requirements.txt
flask==3.0.0
firebase-admin==6.4.0
python-dotenv==1.0.0
groq==0.4.2
requests==2.31.0
Werkzeug==3.0.1
Jinja2==3.1.2
gunicorn==21.2.0
